# 47 In Progress (Added save-temps to avm2-as tool, added new packages, updated docs)

+++ tools/as/as.cpp
@@ -288,7 +288,7 @@ int main(int argc, char **argv)
   static const char *defOutput = "a.out";
   // TODO: option for verbose mode, logging (VPMedia)
   bool targetPlayer = false;
-  //bool useLegacyAsc = false;
+  bool saveTemps = false;
   FILE *input = NULL;
   FILE *output = NULL;
 
@@ -311,10 +311,14 @@ int main(int argc, char **argv)
         error("Multiple input files specified");
       input = stdin;
     }
-    else if(arg == "--target-player")
+    else if(arg == "--target-player") 
+    {
       targetPlayer = true;
-    //else if(arg == "--use-legacy-asc")
-    //  useLegacyAsc = true;
+    }
+    else if(arg == "--save-temps") 
+    {
+      saveTemps = true;
+    }
     else
     {
       if(input)
@@ -401,25 +405,16 @@ int main(int argc, char **argv)
   jargs.push_back("-Xms512M");
   jargs.push_back("-Xmx2048M");
   jargs.push_back("-ea");
-  //if(useLegacyAsc)
-  //{
-  //  jargs.push_back("-classpath");
-  //  jargs.push_back(libPath + "/asc.jar");
-  //  jargs.push_back("macromedia.asc.embedding.ScriptCompiler");
-  //}
-  //else
-  //{
-    jargs.push_back("-jar");
-    jargs.push_back(libPath + "/asc2.jar");
-    // merge the compiled source into a single output
-    jargs.push_back("-merge");
-    // emit metadata information into the bytecode
-    jargs.push_back("-md");
-    // turn on parallel generation of method bodies feature
-    jargs.push_back("-parallel");
-    // turn on the inlining of functions
-    jargs.push_back("-inline");
-  //}
+  jargs.push_back("-jar");
+  jargs.push_back(libPath + "/asc2.jar");
+  // merge the compiled source into a single output
+  jargs.push_back("-merge");
+  // emit metadata information into the bytecode
+  jargs.push_back("-md");
+  // turn on parallel generation of method bodies feature
+  jargs.push_back("-parallel");
+  // turn on the inlining of functions
+  jargs.push_back("-inline");
   // future abc (?)
   jargs.push_back("-abcfuture");
   // use the AS3 class based object model
@@ -453,13 +448,8 @@ int main(int argc, char **argv)
   jargs.push_back("-import");
   jargs.push_back(libPath + "/CModule.abc");
 
-  //if(useLegacyAsc) {
-  //  jargs.push_back(unipath(tmp1Path));
-  //  jargs.push_back(unipath(tmp2Path));
-  //} else {
-    jargs.push_back(unipath(tmp2Path));
-    jargs.push_back(unipath(tmp1Path));
-  //}
+  jargs.push_back(unipath(tmp2Path));
+  jargs.push_back(unipath(tmp1Path));
 
   jargs.push_back("-outdir");
   jargs.push_back(unipath(dirname(outTmpPath)));
@@ -492,7 +482,8 @@ int main(int argc, char **argv)
   fclose(outTmp);
 
   #ifdef _WIN32
-    DeleteFile(apptempdir);
+    if(!saveTemps)
+        DeleteFile(apptempdir);
   #endif
   
   return 0;
